Test terminados, 99% de cobertura. Falta diagrama

+++ src/main/java/Solitario/Mazo.java
@@ -7,13 +7,11 @@ public abstract class Mazo{
     public static final int CARTAS_POR_PALO = 13;
     protected Deque<Carta> mazo;
 
-    public Mazo(){
-        this.mazo = new LinkedList<>();
-    }
+    public Mazo(){this.mazo = new LinkedList<>();}
 
     public abstract void inicializar();
 
-    public void mezclar(int semilla){
+    public void mezclar(long semilla){
         List<Carta> aux = new ArrayList<>(mazo);
         Collections.shuffle(aux, new Random(semilla));
         mazo = new LinkedList<>(aux);
